Add the code to generate the grid
Flexboxes are used to generate the grid.

As of now, the number of rows and columns of the grid are set on the
.js file. In a future version, this number will be retrieved directly
from the website interface.

The grid adapts to the available space, taking into account the aspect
ratio of the grid and of the available space.

As of now, the grid elements might overflow the grid container when
the page is resized or the orientation changes.

Some events listeners needs to be added in order to adapt the grid
when the page is resized or the orientation changes.