Add toggle slider to keep the grid aspect ratio when modifying it
A toggle slider has been added to keep the grid aspect ratio when
either the number of rows or columns of the grid are modified.

Note that as the elements of the grid are squares, such aspect
ratio is not always the actual final one. In particular, when
the number of columns is fixed, the number of rows is to be
computed, and vice versa, and some rounding is applied.