Add random color mode functionality
Random color mode can be enabled, by which a random color is generated in
each cell or at each click.

Global variable penColor is now an anonymous function, whose definition
depends on the current mode. In 'one color mode', this is simply the anonymous
function ()=>color, while in 'random color mode', it returns a random color,
where the H, S and L components are generated within a given range.

ColorSetOnEnterOrDown property is added to each cell, which is set to true
when a color is permanent applied to a cell (after a pointerenter or
pointerdown event), and restored to false when the pointer leaves the cell.
This is used to impose the following behaviour:
- touch: at each tap or when sliding in, a new color is generated, and
         permanently applied.
- mouse: a new color is generated, and permanently applied at each right
         click on the cell, provided that a new color has already been
         permanentely applied since the mouse entered the cell. Otherwise,
         the the background color is permanently applied.

Note that, from now on, the elements of the grid are referred to as 'cells'.