Add selectors to selects the ranges to generate random colors
It is assumed that random colors are generated by randomly generating
the H (hue), S (saturation) and L (lightness) components.

Three range selectors are added to choose the range for such components.
To select a range, two sliders are needed in the same interval, to get a
'minimum' (from slider 1) and a 'maximum' (from slider 2) value for the
component.

As the standard <input> tag does not support double sliders, the
noUIslider library has been used and customized. noUIslider is indeed
a lightweight JavaScript range slider with full touch support.

Note that the Hue component is a circular data, as it is an angle.
Indeed, the angle 0° corresponds to the angle 360° and multiples.
Then, one should be able also to select intervals like, e.g., [270°,10°],
which are actually to be interpreted as [270°,360°+10°]. In this case,
slider 1 still sets the minimum, while the value from slider 2 must be
increased by the value of 360° to get the maximum. This is to be done
every time the value from slider 2 is smaller than the value from slider
1. In that case, the acutual interval for valid data is the union of the
outer intervals [270°,359°]U[0°,10°].

Some code cleaning has also been done in the .html, .css, and .js files.