Add fade mode functionality between two colors
It is now possible to select the fade mode, which works as follows.
When the user moves the cursor keeping the left button pushed, or
a touch slides occurs, the printed color changes at each cell. The
colors are selected according to a linear gradient between two colors
(fromFadeColor, toFadeColor), with a predefined number of steps (equal
to 1/fracFade).
The gradient color is reset at each tap or when clicking in a new cell.

fromFadeColor, toFadeColor and fracFade are hardcoded in this first version.

In this mode, multitouch is currently not tested, but should be broken.
Tests needs to be performed.