Add support for multi-touch in fade mode
In fade mode, having a single global variable currentFadeIndex implies that, when multi-touch
is supported, when sliding occurs simultaneously in multiple points, currentFadeIndex is
incremented without taking into account those points. As a result, some steps in the gradient
are skipped while sliding.

To fix this, instead of having a single global variable currentFadeIndex, a global list of
variables currentFadeIndexList is introduced, whose elements refer to different pointers.
Recall that given a pointer event, e.g., a touch followed by a slide, a unique id is
assigned to it.

currentFadeIndexList is implemented as an object whose indices are indeed the ids of such pointers
(retrieved from event.PointerId, within the pointer events callbacks).

currentFadeIndexList is reset to an empty object each time the mode changes or changes occurs to
the fade settings, to limit the size currentFadeIndexList.