Fix touch behaviour in fading mode
In fading mode, when using a mouse, successive clicking in the same
cell increases its color with respect to the selected gradient.
Before this fix, this did not happened when using touch. This is
because a pointerleave event is fired at every touch tap inside the
same cell (while this does not happen at mouse taps in the same cell).
Then, a new pointer ID is used for the next touch tap, so a new
element in currentFadeIndexList is initialized to 0.

Now, downCell global variable is used to test whether a touch tap
occurs at the same cell as before or not. In the former case, the
fade index for tapping events (stored in the global variable
downIndex) is restored.

Multitouch is not supported for this feature.